+++ sff/strings.py
@@ -1,5 +1,5 @@
 # owner: Midrag
-VERSION = "4.6.2"
+VERSION = "4.6.3"
 STEAM_WEB_API_KEY = "1DD0450A99F573693CD031EBB160907D"
 """NOTE: Not even mine lmao. Given away by oureveryday
 (source: https://github.com/SteamAutoCracks/Steam-auto-crack/issues/33)"""